pyproject.toml
[project]
name = "syndesi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adam-atan2-pytorch>=0.2.4",
    "causal-conv1d>=1.5.3.post1",
    "datasets>=3.1.0",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "lightning>=2.3.3",
    "litmodels>=0.1.8",
    "mamba-ssm>=2.2.6.post3",
    "rdkit>=2025.9.1",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torch>=2.5.1",
    "transformers>=4.57.1",
]
